requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pygame-ce>=2.4.0

